backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
google-analytics-data==0.17.1
google-generativeai==0.3.5
httpx==0.25.2
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
apscheduler==3.10.4
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-httpx==0.26.0
python-multipart==0.0.6
